requirements.txt
python-dotenv==1.1.0
paramiko==3.5.1
scp==0.15.0
requests==2.32.3
pika==1.3.2
pytz==2025.2